+++ src/main/scala/sc4pac/Constants.scala
@@ -35,7 +35,7 @@ object Constants {
     * - 7: conditional variants
     * - 8: authentication (#48)
     */
-  val channelSchemeVersions: Range = 1 to 8  // supported versions
+  val channelSchemeVersions: Range = 6 to 8  // supported versions
 
   val pluginsLockScheme = 2
   val bufferSizeExtract = 64 * 1024  // 64 kiB, bounded by disk speed